Add preload generation
